MGDOBR-590: move msw stuff to dedicated folder
