+++ mocked-api/browser.ts
