+++ tsconfig.json
@@ -6,6 +6,7 @@
     "types",
     "webpack.config.js",
     "__mocks__",
+    "mocked-api",
     "setupTests.ts"
   ],
   "compilerOptions": {
